helpers/requirements.txt
staticmap==0.5.5
Pillow==8.1.1
